fixing ovenarea / round 2
also updated unit tests so that it runs the command validation logic of decisions
